Move path around
