+++ python/tvm/relay/__init__.py
@@ -64,13 +64,6 @@ from .scope_builder import ScopeBuilder
 # Required to traverse large programs
 setrecursionlimit(10000)
 
-__STD_PATH__ = os.path.join(os.path.dirname(os.path.realpath(__file__)), "std")
-
-@register_func("tvm.relay.std_path")
-def _std_path():
-    global __STD_PATH__
-    return __STD_PATH__
-
 # Span
 Span = base.Span
